fix: AboutMeの修正を3点行った①タグを修正した②不要な記述を削除した③コンポーネントを分割した
